requirements.txt
tensorboard==2.10.0
accelerate==0.14.0
-e git+https://github.com/openai/CLIP.git@main#egg=clip
diffusers==0.14.0
pillow==9.3.0
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
transformers==4.25.1
wandb==0.13.5
clip_retrieval==2.36.1
git+https://github.com/cloneofsimo/lora.git